+++ tests/test_plotters_combinedplotter_icon_2d.py
@@ -1,0 +1,22 @@
+"""Test module of the :mod:`psyplot.plotter.maps` module"""
+import unittest
+import test_base as tb
+import test_plotters_combinedplotter as tpc
+import test_plotters_combinedplotter_icon as tpci
+
+
+class IconCombinedPlotterTest2D(tb.TestBase2D, tpci.IconCombinedPlotterTest):
+    """Test :class:`psyplot.plotter.maps.CombinedPlotter` class for icon grid
+    without time and vertical dimension"""
+
+    var = ['t2m', ['u_2d', 'v_2d']]
+
+    def _label_test(self, key, label_func, has_time=None):
+        if has_time is None:
+            has_time = not bool(self.vector_mode)
+        tpc.CombinedPlotterTest._label_test(
+            self, key, label_func, has_time=has_time)
+
+
+if __name__ == '__main__':
+    unittest.main()
